Build protobuf
