+++ src/project_config.rs
@@ -1,0 +1,1 @@
+include!(concat!(env!("OUT_DIR"), "/generated/mod.rs"));
